Added Makefile
I added a Makefile to simplify compilation using G++. This is better practice than including the source code files at the end of the header files and simpler to use, as you can just run the `make` command in the terminal if Make is installed. `make clean` will delete the executable.