Metadata included in block file
I've added metadata to the blocked data file.
This is in the format: LI,RBN,#ofRecords,prevBlock,nextBlock,records...

where LI is the length-indicator of the metadata, including the LI itself, RBN is the relative block number, prevBlock is the RBN is of the previous block, and nextBlock is the RBN of the subsequent block.

If there is no next or previous blocks, then those fields are -1.