Move file opening to main program from buffers
I moved the file opening to the main program so that it instead passes files by reference to the buffers. This was necessary to initialize ZipCodeBuffer and BlockBuffer with the same file.

BlockBuffer is now initialized with an ifstream rather than a string.

I started some preliminary code to handle reading blocks. It still needs a lot of work.