Add core block reading functionality
ZipCodeBuffer now uses BlockBuffer to read and unpack the metadata and records in each block. BlockBuffer uses the RBN links to read the blocks in logical order. It can read the file and display the table the same way as the other file types.

The header buffer is not yet implemented, but it is built with it in mind. The index and search do not yet work with the new file type.

Since us_postal_codes.txt now has the updated header record, which the program does not yet support, I created a new us_postal_codes_simple.txt file for testing, which has the old header. blockedcodes.txt is the blocked file with the old header (and is missing the last block due to the converter having a bug, so I manually edited the second to last block to be the last one).